Add explicit permissions in the pipeline
We need this now as we're increasing security measures in the organization.

I modified the pipeline a bit to remove sed regex where we could do just okay with node.

I also put the env vars up into job context, so that we don't have to repeat it multiple times across the pipeline.